requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
websockets==14.2
twilio==9.2.3
supabase==2.15.0
openai==1.76.0
python-dotenv==1.0.1
python-multipart==0.0.20
stripe==6.6.0
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dateutil==2.8.2
